1. CSVs are now saved for AllTransactions, AllBudgets, and AllAccounts 2. Code reorganization 3. Converted all float to double, to avoid precision issues faced

+++ iSaveMoneyBackupProcessor/DataModels/Account.cs
@@ -5,7 +5,7 @@
         public int id { get; set; }
         public string name { get; set; }
         public int type { get; set; }
-        public float balance { get; set; }
+        public double balance { get; set; }
         public bool active { get; set; }
     }
 }
